basic changes applied
