+++ lib/greet.js
