+++ lib/singleLine.js
@@ -1,5 +1,6 @@
 
-var greet = require('./greet');
+function singleLine() {
+    var greet = require('./greet');
 
     var allParts = '';
     process.argv.forEach((val, idx, arr) => {
@@ -9,9 +10,9 @@ var greet = require('./greet');
         return allParts;
     });
 
-    var anotherGreeting = greet(allParts);
-    console.log(anotherGreeting);
-
+    var anotherGreeting = greet(allParts).trim(); // get rid of the trailing space
     return anotherGreeting;
+};
 
-// module.exports = singleLine;
+console.log(singleLine());
+module.exports = singleLine;
